Rename `Link` component to `TextLink` and pass in the display value via `label` prop instead of `children`
