+++ src/lib/components/Button/README.mdx
@@ -399,7 +399,7 @@ animation is made.
 
 In addition to the options below, you can specify [React synthetic events] or
 any custom HTML attributes that do not interfere with the API, and they will be
-passed to the `button` HTML element. This enables making the component
+passed to the `<button>` HTML element. This enables making the component
 interactive and helps improve its accessibility.
 
 <Props table of={Button} />
