--- src/lib/components/Link/Link.scss
@@ -1,11 +1,0 @@
-@use "theme";
-
-.root {
-    text-decoration: theme.$link-text-decoration;
-    color: theme.$link-color;
-
-    &:hover {
-        text-decoration: theme.$link-hover-text-decoration;
-        color: theme.$link-hover-color;
-    }
-}
