--- src/lib/components/Link/README.mdx
@@ -1,85 +1,0 @@
----
-name: Link
-menu: 'Actions'
-route: /components/link
----
-
-# Link
-
-import {
-  Playground,
-  Props,
-} from 'docz'
-import Placeholder from '../../../docs/_components/Placeholder'
-import { Link } from './Link'
-
-Link allow users to follow navigation.
-
-## Basic Usage
-
-To implement the Link component, you need to import it first:
-
-```js
-import { Link } from '@react-ui-org/react-ui';
-```
-
-And use it:
-
-<Playground>
-  <Link href="/contribute/guidelines">How can I contribute to React UI?</Link>
-</Playground>
-
-See [API](#api) for all available options.
-
-## General Guidelines
-
-**Avoid using links for actions**, use a [Button](/components/button)
-instead. This is because users expect navigation to happen when clicking on
-something what looks like a link.
-
-## Block-Level Content
-
-Besides text, link can contain block-level elements too.
-
-<Playground>
-  <Link href="/contribute/guidelines">
-    <Placeholder>Block-level element wrapped with link</Placeholder>
-  </Link>
-</Playground>
-
-## Custom Routing
-
-It's common to use custom function for routing within SPAs. Use the
-`onClick` option to provide such function.
-
-<Playground>
-  <Link
-    href="/contribute/guidelines"
-    onClick={(event) => {
-      event.preventDefault();
-      window.location = event.currentTarget.getAttribute('href');
-    }}
-  >
-    This link is controlled by custom function
-  </Link>
-</Playground>
-
-## API
-
-In addition to the options below, you can specify [React synthetic events] or
-any custom HTML attributes that do not interfere with the API, and they will be
-passed to the `a` HTML element. This enables making the component interactive
-and helps improve its accessibility.
-
-<Props table of={Link} />
-
-## Theming
-
-| Custom Property                      | Description                                  |
-|--------------------------------------|----------------------------------------------|
-| `--rui-Link__color`                  | Text color                                   |
-| `--rui-Link__text-decoration`        | Text decoration, e.g. underline              |
-| `--rui-Link--hover__color`           | Text color on hover                          |
-| `--rui-Link--hover__text-decoration` | Text decoration on hover                     |
-
-[React synthetic events]: https://reactjs.org/docs/events.html
