--- src/lib/components/Link/_theme.scss
@@ -1,4 +1,0 @@
-$link-color: var(--rui-Link__color);
-$link-text-decoration: var(--rui-Link__text-decoration);
-$link-hover-color: var(--rui-Link--hover__color);
-$link-hover-text-decoration: var(--rui-Link--hover__text-decoration);
