--- src/lib/components/Link/index.js
@@ -1,1 +1,0 @@
-export { default } from './Link';
