+++ src/lib/components/Radio/README.mdx
@@ -287,7 +287,7 @@ It's possible to disable just some options or the whole set.
 
 In addition to the options below, you can specify [React synthetic events] or
 any custom HTML attributes that do not interfere with the API, and they will be
-passed to the `input` HTML element. This enables making the component
+passed to the `<input>` HTML element. This enables making the component
 interactive and helps improve its accessibility.
 
 <Props table of={Radio} />
