+++ src/lib/components/TextArea/README.mdx
@@ -350,7 +350,7 @@ It's possible to disable the whole input.
 
 In addition to the options below, you can specify [React synthetic events] or
 any custom HTML attributes that do not interfere with the API, and they will be
-passed to the `textarea` HTML element. This enables making the component
+passed to the `<textarea>` HTML element. This enables making the component
 interactive and helps improve its [accessibility](#forwarding-html-attributes).
 
 <Props table of={TextArea} />
