+++ src/lib/components/TextLink/_theme.scss
@@ -1,0 +1,4 @@
+$color: var(--rui-TextLink__color);
+$text-decoration: var(--rui-TextLink__text-decoration);
+$hover-color: var(--rui-TextLink--hover__color);
+$hover-text-decoration: var(--rui-TextLink--hover__text-decoration);
