+++ src/lib/theme.scss
@@ -863,13 +863,13 @@
     --rui-FormField--check--checked__check-background-color: var(--rui-color-active);
 
     //
-    // Link
-    // ====
+    // TextLink
+    // ========
 
-    --rui-Link__color: var(--rui-link-color);
-    --rui-Link__text-decoration: var(--rui-link-decoration);
-    --rui-Link--hover__color: var(--rui-link-hover-color);
-    --rui-Link--hover__text-decoration: var(--rui-link-hover-decoration);
+    --rui-TextLink__color: var(--rui-link-color);
+    --rui-TextLink__text-decoration: var(--rui-link-decoration);
+    --rui-TextLink--hover__color: var(--rui-link-hover-color);
+    --rui-TextLink--hover__text-decoration: var(--rui-link-hover-decoration);
 
     //
     // Modal
